generate emoji images without md5 digest
Sprockets seems to generate every image with digests now.
However, in emojify.js (which does not belong to us but to
the upstream repository) we do not (and don't want to) use
image-url('..') or similar things to generate the actual
digested path of each emoji image.

Instead we use the non-stupid-digest-assets gem (which
is already required by OpenProject) to whitelist our
images from digestion :)

fixes #3